+++ packages/axios-extension/src/auth/reentrance-ticket/index.ts
@@ -2,13 +2,9 @@ import type { Logger } from '@sap-ux/logger';
 import type { AddressInfo } from 'node:net';
 import open = require('open');
 import { defaultTimeout } from '../connection';
-import { ABAPSystem } from './abap-system';
+import { ABAPVirtualHostProvider } from './abap-virtual-host-provider';
 import { setupRedirectHandling } from './redirect';
 
-/**
- * DO NOT USE THIS SERVICE ENDPOINT DIRECTLY.
- * It might be removed in the future without notice.
- */
 const ADT_REENTRANCE_ENDPOINT = '/sap/bc/sec/reentrance';
 
 /**
@@ -27,19 +23,21 @@ export async function getReentranceTicket({
     backendUrl: string;
     logger: Logger;
     timeout?: number;
-}): Promise<{ reentranceTicket: string; apiUrl?: string }> {
+}): Promise<{ reentranceTicket: string; backend?: ABAPVirtualHostProvider }> {
+    const backend = new ABAPVirtualHostProvider(backendUrl, logger);
+    const uiHostname = await backend.uiHostname();
     return new Promise((resolve, reject) => {
-        const backend = new ABAPSystem(backendUrl);
         // Start local server to listen to redirect call, with timeout
         const { server, redirectUrl } = setupRedirectHandling({ resolve, reject, timeout, backend, logger });
         server.listen();
-
         const redirectPort = (server.address() as AddressInfo).port;
 
         // Open browser to handle SAML flow and return the reentrance ticket
         const scenario = process.env.FIORI_TOOLS_SCENARIO ?? 'FTO1';
         const endpoint = process.env.FIORI_TOOLS_REENTRANCE_ENDPOINT ?? ADT_REENTRANCE_ENDPOINT;
-        const url = `${backend.uiHostname()}${endpoint}?scenario=${scenario}&redirect-url=${redirectUrl(redirectPort)}`;
-        open(url)?.catch((error) => logger.error(error));
+        const url = `${uiHostname}${endpoint}?scenario=${scenario}&redirect-url=${redirectUrl(redirectPort)}`;
+
+        const result = open(url)?.catch((error) => logger.error(error));
+        return result;
     });
 }