+++ packages/ui5-application-inquirer/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @sap-ux/ui5-application-inquirer
 
+## 0.15.19
+
+### Patch Changes
+
+-   @sap-ux/inquirer-common@0.7.51
+
+## 0.15.18
+
+### Patch Changes
+
+-   Updated dependencies [247a5a9]
+    -   @sap-ux/project-input-validator@0.6.26
+
 ## 0.15.17
 
 ### Patch Changes