+++ packages/ui5-library-reference-sub-generator/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @sap-ux/ui5-library-reference-sub-generator
 
+## 0.1.52
+
+### Patch Changes
+
+-   Updated dependencies [bacaf93]
+    -   @sap-ux/fiori-generator-shared@0.13.23
+    -   @sap-ux/feature-toggle@0.3.2
+    -   @sap-ux/telemetry@0.6.29
+    -   @sap-ux/ui5-library-reference-inquirer@0.4.52
+
 ## 0.1.51
 
 ### Patch Changes